requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pypdf
python-docx
python-pptx
requests